==> R/01_load.qmd <==
---
title: "Load data"
format: html
editor: source
author: authors
---

### Source functions

```{r}
source("../R/99_proj_func.R")
```

### Libraries:

```{r}
#| message: false
library("tidyverse")
library(httr)
library(jsonlite)
library(dplyr)
library(R.utils)
library(illuminaio)
```

### Generate data directories

```{r}
make_dir("../_raw")
make_dir("../data")
```

### Load files:

```{r}
#| message: false

raw_dir <- "../_raw/"
download_dataset_ncbi(raw_dir)
download_data_annotation_ncbi(raw_dir)

ncbi_data <- read_unstructured_ncbi_table(str_c(raw_dir, "raw_ncbi_data.txt.gz"))
annotation_data <- read_bgx_file(raw_dir, "raw_annotation.bgx.gz")

ncbi_data |> sample_n(5)

annotation_data
annotation_data$probes
```

### Write files

We should write a tsv file to our data folder containing the data frames that have been loaded in. To be improved...

```{r}
write_tsv(ncbi_data,
            file = "../data/01_ncbi_data.tsv.gz")

write_tsv(annotation_data$probes,
            file = "../data/01_ncbi_annot_probes.tsv.gz")
```
