==> R/03_describe.qmd <==
### Source functions

```{r}
source("../R/99_proj_func.R")
```

### Libraries:

```{r}
#| message: false
library("tidyverse")
library("table1")
library("patchwork")
```

### Loading data:

```{r}
#| eval: true
#| message: false
pheno_data <- vroom::vroom("../data/02_ncbi_clean_pheno.tsv.gz", delim = "\t")
expr_data <- vroom::vroom("../data/02_transposed_expr_data.tsv.gz", delim = "\t")

pheno_data |> sample_n(5)
expr_data |> sample_n(5)
```

### Study subjects description

```{r}
num_mothers = 65

pheno_data |>
  slice_head(n = num_mothers) |> 
  table1(x = formula(~ age + maternal_bmi + gestational_age + newborn_weight | smoking_status),
         data = _)

```

### Some overall plots:

```{r}
bp_age <- pheno_data |>
  ggplot(aes(y = age,
             fill = smoking_status)) +
  geom_boxplot() +
  labs(y = "Mother age (years)",
       x= "",
       caption = "",
       title = "Mother's age",
       fill = "Smoking Status") +
  theme(legend.position = "bottom",
        legend.direction = "vertical",
        legend.title = element_text(size=10),
        legend.text = element_text(size=8),
        plot.margin=grid::unit(c(0,0,0,0), "mm"))  +
  theme_minimal() + 
  scale_fill_manual(values = c("gold2","red"))
ggsave("../results/plots/description_age_boxplot.png")

bp_bmi <- pheno_data |>
  ggplot(aes(y = maternal_bmi,
             fill=smoking_status)) +
  geom_boxplot() +
  labs(x = "",
       y= "Maternal BMI",
       caption = "",
       title = "Mother's BMI",
       fill = "Smoking Status") +
  theme(legend.position = "bottom",
        legend.direction = "vertical",
        legend.title = element_text(size=10),
        legend.text = element_text(size=8),
        plot.margin=grid::unit(c(0,0,0,0), "mm"))  +
  coord_cartesian(ylim = c(15,
                           40)) +
  theme_minimal() +
  scale_fill_manual(values = c("gold2","red"))
ggsave("../results/plots/description_BMI_boxplot.png")

bp_maternal_cotinine <- pheno_data |>
  ggplot(aes(y = cotinine_conc_maternal,
             fill = smoking_status)) +
  geom_boxplot() +
  labs(x = "",
       y= "Blood cotinine (ng/ml)",
       caption = "",
       title = "Maternal blood Cotinine Concentration",
       fill = "Smoking Status") +
  theme(legend.position = "bottom",
        legend.direction = "vertical",
        legend.title = element_text(size=10),
        legend.text = element_text(size=8),
        plot.margin=grid::unit(c(0,0,0,0), "mm"))  +
  coord_cartesian(ylim = c(0,
                           50)) +
  theme_minimal() +
  scale_fill_manual(values = c("gold2","red"))
ggsave("../results/plots/description_bloodCotinine_boxplot.png")

bp_cord_cotinine <- pheno_data |>
  ggplot(aes(y = cotinine_conc_cord,
             fill = smoking_status)) +
  geom_boxplot() +
  labs(x = "",
       y= "Cotinine (ng/ml)",
       caption = "",
       title = "Cord Blood Cotinine Concentration",
       fill = "Smoking Status") +
  theme(legend.position = "bottom",
        legend.direction = "vertical",
        legend.title = element_text(size=10),
        legend.text = element_text(size=8),
        plot.margin=grid::unit(c(t = -20, r = 0, b = -20, l = 0), "mm")) +
  coord_cartesian(ylim = c(0,
                           50)) +
  theme_minimal() +
  scale_fill_manual(values = c("gold2","red"))
ggsave("../results/plots/description_CordCotinine_boxplot.png")

bp_newbornweight <- pheno_data |>
  ggplot(aes(y = newborn_weight,
             fill = smoking_status)) +
  geom_boxplot() +
  labs(x = "",
       y= "Weigth (g)",
       caption = "",
       title = "Newborn Weight",
       fill = "Smoking Status") +
  theme(legend.position = "bottom",
        legend.direction = "vertical",
        legend.title = element_text(size=10),
        legend.text = element_text(size=8),
        plot.margin=grid::unit(c(t = -20, r = 0, b = -20, l = 0), "mm")) +
  coord_cartesian(ylim = c(2500,
                           5500)) +
  theme_minimal() +
  scale_fill_manual(values = c("gold2","red"))
ggsave("../results/plots/description_newbornWeight_boxplot.png")


bp_placentalweight <- pheno_data |>
  ggplot(aes(y = placental_weight,
             fill = smoking_status)) +
  geom_boxplot() +
  labs(x = "",
       y= "Weigth (g)",
       caption = "",
       title = "Placental Weight",
       fill = "Smoking Status") +
  theme(legend.position = "bottom",
        legend.direction = "vertical",
        legend.title = element_text(size=10),
        legend.text = element_text(size=8),
        plot.margin=grid::unit(c(t = -20, r = 0, b = -20, l = 0), "mm"))  +
  coord_cartesian(ylim = c(450,
                           700)) +
  theme_minimal() +
  scale_fill_manual(values = c("gold2","red"))
ggsave("../results/plots/description_placentaWeight_boxplot.png")


combined_plot1 <- (
  bp_placentalweight + theme(legend.position = "none") + 
     (bp_newbornweight + theme(legend.position = "none")
      )) / bp_cord_cotinine + 
     
     plot_annotation(title = "Comparison by Smoking Status",
                     caption = "Data from NCBI (2011)",
                     tag_levels = 'A',
                     tag_prefix = 'Fig. ',
                     tag_sep = '.', 
                     tag_suffix = ':',
                     theme = theme(legend.position = "topright",
                                   plot.title = element_text(size = 18,
                                                             hjust = 0.5, 
                                                             face = "bold", 
                                                             margin = margin(0,0,0,0),
                                                             color = "purple4"),
                                   plot.margin=grid::unit(c(t = 0, r = 0, b = 0, l = -1.3), "mm"))) +
  plot_layout(guides = "collect", heights = unit(c(1.35, 2.35), c('cm', 'cm')))
ggsave("../results/plots/description_combined_plot1.png")
combined_plot1


```
