==> R/04_augment.qmd <==
---
title: "Augment data"
format: html
editor: source
author: authors
---

### Source functions

```{r}
source("../R/99_proj_func.R")
```

### Library

```{r}
#| eval: true
#| echo: true

library("tidyverse")
```

### Loading

```{r}
#| eval: true
#| echo: true

# load tsv files
ncbi_clean_pheno <- vroom::vroom("../data/02_ncbi_clean_pheno.tsv.gz",
                                 delim = "\t")
transposed_expr_data <- vroom::vroom("../data/02_transposed_expr_data.tsv.gz", 
                            delim = "\t")
transposed_placenta_expr <- vroom::vroom("../data/02_transposed_placenta_expr.tsv.gz", 
                            delim = "\t")
transposed_maternal_expr <- vroom::vroom("../data/02_transposed_maternal_expr.tsv.gz", 
                            delim = "\t")
transposed_cord_expr <- vroom::vroom("../data/02_transposed_cord_expr.tsv.gz", 
                            delim = "\t")
```

### BMI class and Age group augmenation:

```{r}
#| eval: true
#| echo: true

ncbi_augment_pheno <- ncbi_clean_pheno |>
  drop_na(maternal_bmi) |>
  mutate(BMI_class = case_when(maternal_bmi < 16.5 ~ "Severely underweight",
                               16.5 <= maternal_bmi & maternal_bmi < 18.5 ~ "Underweight",
                               18.5 <= maternal_bmi & maternal_bmi < 24.9 ~ "Normal weight",
                               24.9 <= maternal_bmi & maternal_bmi < 30 ~ "Overweight",
                               maternal_bmi >= 30 & maternal_bmi < 35 ~ "Obesity class I",
                               maternal_bmi >= 35 & maternal_bmi < 40 ~ "Obesity class II",
                               maternal_bmi >= 40 ~ "Obesity class III")) |>
  mutate(BMI_class = factor(BMI_class,
                            levels = c("Severely underweight",
                                       "Underweight",
                                       "Normal weight", 
                                       "Overweight",
                                       "Obesity", 
                                       "Obesity class I",
                                       "Obesity class II", 
                                       "Obesity class III"))) |>
  relocate(BMI_class,
           .before = maternal_bmi) |>
  select(-maternal_bmi) |>
  mutate(age_group = cut(age,
                         breaks = c(10,
                                    20,
                                    30,
                                    40,
                                    50))) |>
  relocate(age_group,
           .before = age)
```

### Joining and creating log fold change dataframes

```{r}

# join transposed expr data with phenotypical data
ncbi_joined <- ncbi_augment_pheno |>
  full_join(transposed_expr_data)

placenta_joined <- ncbi_augment_pheno |>
  inner_join(transposed_placenta_expr)

maternal_joined <- ncbi_augment_pheno |>
  inner_join(transposed_maternal_expr)

cord_joined <- ncbi_augment_pheno |>
  inner_join(transposed_cord_expr)

# calculate log fold change for all data frames
smoking_fc_placenta <- smoking_expr_fc(placenta_joined)

smoking_fc_maternal <- smoking_expr_fc(maternal_joined)

smoking_fc_cord <- smoking_expr_fc(cord_joined)

smoking_fc_all <- smoking_expr_fc(ncbi_joined)

```

### Write tsv file

```{r}
#| eval: true
#| echo: true

write_tsv(x = ncbi_joined,
          file = "../data/04_ncbi_joined.tsv.gz")

write_tsv(x = placenta_joined,
          file = "../data/04_placenta_joined.tsv.gz")

write_tsv(x = maternal_joined,
          file = "../data/04_maternal_joined.tsv.gz")

write_tsv(x = cord_joined,
          file = "../data/04_cord_joined.tsv.gz")

write_tsv(x = smoking_fc_all,
          file = "../data/04_smoking_fc_all.tsv.gz")

write_tsv(x = smoking_fc_placenta,
          file = "../data/04_smoking_fc_placenta.tsv.gz")

write_tsv(x = smoking_fc_maternal,
          file = "../data/04_smoking_fc_maternal.tsv.gz")

write_tsv(x = smoking_fc_cord,
          file = "../data/04_smoking_fc_cord.tsv.gz")
```
