==> R/05_analysis_1.qmd <==
### Source functions

```{r}
source("../R/99_proj_func.R")
```

### Library:

```{r}
#| message: false
library("tidyverse")
library("ggridges")
library("viridis")
library("broom")
library("ggrepel")
library("corrplot")
library("plyr")
library("patchwork")
library("msigdbr")
library("fgsea")
```

### loading:

```{r}
#| eval: true
#| message: false

ncbi_joined <- vroom::vroom("../data/04_ncbi_joined.tsv.gz",
                             delim = "\t")

placenta_joined <- vroom::vroom("../data/04_placenta_joined.tsv.gz",
                                delim = "\t")

maternal_joined <- vroom::vroom("../data/04_maternal_joined.tsv.gz",
                                delim = "\t")

cord_joined <- vroom::vroom("../data/04_cord_joined.tsv.gz",
                                delim = "\t")

smoking_fc_cord <- vroom::vroom("../data/04_smoking_fc_cord.tsv.gz",
                                 delim = "\t")

smoking_fc_placenta <- vroom::vroom("../data/04_smoking_fc_placenta.tsv.gz",
                                 delim = "\t")


ncbi_joined |> 
  sample_n(5)
placenta_joined |> 
  sample_n(5)
```

### Fitting model to all sample types:

```{r}
regression_placenta <- create_models(placenta_joined)

regression_cord <- create_models(cord_joined)

regression_maternal <- create_models(maternal_joined)

```

### Volcano plot:

```{r}
#| warning: false

significant_placenta <- subset(regression_placenta , 
                          is_significant == "yes") |>
  dplyr::select(gene)

significant_cord <- subset(regression_cord , 
                          is_significant == "yes") |>
  dplyr::select(gene)

significant_maternal <- subset(regression_maternal , 
                          is_significant == "yes") |>
  dplyr::select(gene)

regression_placenta <- regression_placenta |>
  mutate(sample = "placenta")

regression_maternal <- regression_maternal |>
  mutate(sample = "maternal")

regression_cord <- regression_cord |>
  mutate(sample = "cord")

regression_all <- join_all(list(regression_placenta,
                                regression_maternal,
                                regression_cord),
                           type='full')

```

```{r}
regression_all |>
  ggplot(aes(x = gene,
             y = -log(p.value,10),
             label = ifelse(is_significant == "yes", 
                            gene, 
                            ""))) + 
  geom_point(aes(x = estimate, 
                 color = is_significant),
             alpha = .5,
             size = .3) +
  geom_hline(yintercept = 0) +
  theme_minimal(base_size = 10) +
  scale_x_continuous(limits=c(-.4,.4)) + 
  scale_y_continuous(limits=c(0,3.5)) + 
  labs(title = "Genes Associated with Smoking",
       subtitle = "Genes highlighted in green were significant at a level of signifcance of p < .05",
       x = "Estimates",
       y = "-log10(p)",
       caption = "Data from NCBI (2011)") +
  theme(legend.position = "none",
        plot.title = element_text(hjust = .5,
                                  color = "orange3",
                                  margin = margin(10,-10,10,10),
                                  face = "bold"),
        plot.subtitle = element_text(hjust = .5,
                                     face = "italic"),
        axis.title.y = element_text(face = "italic")) +
  facet_grid(~sample,
             labeller = as_labeller(c(placenta = "Placenta", 
                                      maternal = "Maternal blood",
                                      cord = "Cord blood"))) +
  scale_colour_manual(values = c("purple","green4"))

ggsave("../results/plots/triple_vulcano_plot_smoking.png")
```

### Gene association plot:

blood_association_plot \| cord_association_plot \| placenta_association_plot

sample_blood_plot \| sample_cord_plot \| sample_placenta_plot

```{r}
# one plot of all genes and one plot with a sample of 20 or so genes
blood_association_plot <- regression_maternal |>
  filter(is_significant == "yes") |>
  ggplot(aes(y = fct_reorder(gene,
                             estimate))) + 
  geom_point(aes(x = estimate), 
             size = 1.5) +
  geom_errorbarh(aes(xmax = conf.high,
                     xmin = conf.low),
                     height = 1) +
  geom_vline(xintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(legend.position = "none",
        axis.text.y = element_blank(),
        plot.title = element_text(size = 12, 
                                  hjust = 0.0,
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0), 
                                  color = "darkorange1")) +
  labs(x = "",
       y = "",
       title = "Maternal Blood",)

sample_blood_plot <- regression_maternal |>
  filter(is_significant == "yes") |>
  sample_n(20) |>
  ggplot(aes(y = fct_reorder(gene,
                             estimate))) + 
  geom_point(aes(x = estimate), 
             size = 1.5) +
  geom_errorbarh(aes(xmax = conf.high,
                     xmin = conf.low),
                     height = 1) +
  geom_vline(xintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(legend.position = "bottom",
        plot.title = element_text(size = 12, 
                                  hjust = 0.0,
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0), 
                                  color = "darkorange1")) +
  labs(x = "",
       y = "",
       title = "Maternal Blood",)

cord_association_plot <- regression_cord |>
  filter(is_significant == "yes") |>
  ggplot(aes(y = fct_reorder(gene,
                             estimate))) + 
  geom_point(aes(x = estimate), 
             size = 1.5) +
  geom_errorbarh(aes(xmax = conf.high,
                     xmin = conf.low),
                     height = 1) +
  geom_vline(xintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(legend.position = "none",
        axis.text.y = element_blank(),
        plot.title = element_text(size = 12, 
                                  hjust = 0.0,
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0), 
                                  color = "purple1")) +
  labs(x = "Estimates (95%CIs)",
       y = "",
       title = "Cord Blood",)

sample_cord_plot <- regression_cord |>
  filter(is_significant == "yes") |>
  sample_n(20) |>
  ggplot(aes(y = fct_reorder(gene,
                             estimate))) + 
  geom_point(aes(x = estimate), 
             size = 1.5) +
  geom_errorbarh(aes(xmax = conf.high,
                     xmin = conf.low),
                     height = 1) +
  geom_vline(xintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(legend.position = "bottom",
        plot.title = element_text(size = 12, 
                                  hjust = 0.0,
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0), 
                                  color = "purple1")) +
  labs(x = "Estimates (95%CIs)",
       y = "",
       title = "Cord Blood")

placenta_association_plot <- regression_placenta |>
  filter(is_significant == "yes") |>
  ggplot(aes(y = fct_reorder(gene,
                             estimate))) + 
  geom_point(aes(x = estimate), 
             size = 1.5) +
  geom_errorbarh(aes(xmax = conf.high,
                     xmin = conf.low),
                     height = 1) +
  geom_vline(xintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(legend.position = "bottom",
        plot.title = element_text(size = 12, 
                                  hjust = 0.0,
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0), 
                                  color = "chartreuse4")) +
  labs(x = "",
       y = "",
       title = "Placenta",)

sample_placenta_plot <- regression_placenta |>
  filter(is_significant == "yes") |>
  sample_n(20) |>
  ggplot(aes(y = fct_reorder(gene,
                             estimate),color = )) + 
  geom_point(aes(x = estimate), 
             size = 1.5) +
  geom_errorbarh(aes(xmax = conf.high,
                     xmin = conf.low),
                     height = 1) +
  geom_vline(xintercept = 0) +
  theme_minimal(base_size = 10) +
  theme(legend.position = "none",
        plot.title = element_text(size = 12,
                                  hjust = 0.0,
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0), 
                                  color = "chartreuse4")) +
  labs(x = "",
       y = "",
       title = "Placenta")




# All the plots together
(sample_blood_plot | sample_cord_plot | sample_placenta_plot
  ) / (
    blood_association_plot | cord_association_plot | placenta_association_plot) + 
  plot_annotation(title = "Genes Associated with Smoking",
                  caption = "Data from NCBI (2011)",
                  theme = theme(plot.title = element_text(size = 14,
                                                          hjust = 0.5, 
                                                          face = "bold", 
                                                          margin = margin(0,-10,0,0),
                                                          color = "black")))
ggsave("../results/plots/gene_association.png")
```

We can generally see that when a person smokes there seems to be a equal up and downregulation of diffrent genes. it is diffifult to see anything on the first plot so we try and sampling it. ofcourse the genes have a variety of diffrent function, so it makes sense that the up and downregulation is equally divided. but if we looked at a specific type of genes such as immunity, oxidation, metabolitic and so on. It might have a general up or down regulation

### Gene expression of gene C14orf37 plotted against gene expression of other genes

```{r}

placenta_joined |>
  select(smoking_status,
         C14orf37, CYP1A1, ZFY, CYP1A2, SYNPR, SERPINA3) |> 
  pivot_longer(names_to = "variable",
               values_to = "values",
               3:last_col()) |> 
  ggplot(aes(C14orf37,
             values)) + 
  geom_point(aes(color = smoking_status)) + 
  facet_wrap(~ variable, 
             scales = "free_x", 
             ncol = 6) +  
  geom_smooth(method = "lm",
              se = FALSE,
              color = "darkblue") + 
  scale_y_continuous(trans = "log2") + 
  scale_x_continuous(trans = "log2") +
  theme_linedraw() +
  theme(legend.position = c(0.095, 0.81),
        legend.direction = "vertical",
        legend.title = element_text(size=10),
        legend.text = element_text(size=8),
        legend.background = element_rect(linetype = 2, 
                                         size = 0.5, 
                                         colour = 1),
        plot.title = element_text(size = 14, 
                                  hjust = 0.5, 
                                  face = "bold", 
                                  margin = margin(10,-10,10,10),
                                  color = "darkblue")) + 
  labs(x="Gene expression of C14orf37",
       y="Gene expression",
       color = "Smoking Status",
       title = "Gene expression of 6 genes plotted against that of gene C14orf37",
       caption = "Data from NCBI (2011)") +
  scale_colour_manual(values = c("gold2","red"))

ggsave("../results/plots/multiple_gene_expr_correlation_scaterplot_facet.png")

```

### Gene expression correlation

```{r}

genes_exp_selection <- placenta_joined |> 
  select(28:36)

corrplot_gene_expr <- corrplot(cor(genes_exp_selection),
  type = "upper"
)
```

### Log Fold Change

```{r}

p_fc_placenta <- smoking_fc_placenta |>
  arrange(desc(abs(log2fc))) |> 
  slice_head(n=30) |> 
  ggplot(aes(x=fct_reorder(Gene, abs(log2fc)), y=log2fc)) +
  geom_bar(stat="identity", fill  = "chartreuse4", alpha = 0.6) + 
  coord_flip() +
  theme_minimal() +
  theme(legend.position = "none",
        axis.line = element_line(colour = "chartreuse4", 
                      size = 0.5, 
                      linetype = "solid"),
        axis.title.x = element_text(size = 9),
        plot.title = element_text(size = 12, 
                                  hjust = 0.35, 
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0),
                                  color = "chartreuse4"),
        plot.subtitle = element_text(hjust = 0.5),
        plot.margin=grid::unit(c(0,0,0,0), "mm")) +
  labs(title = "Placenta",
       y = "Log2 Fold Change",
       x = "",
       caption = "")
# Maybe add color based on positive or negative log2fc

ggsave("../results/plots/logfc_of_30_genes_affected_by_smoking_placenta.png")

p_fc_cord <- smoking_fc_cord |>
  arrange(desc(abs(log2fc))) |> 
  slice_head(n=30) |> 
  ggplot(aes(x=fct_reorder(Gene, abs(log2fc)), y=log2fc)) +
  geom_bar(stat="identity", fill  = "purple1", alpha = 0.6) + 
  coord_flip() +
  theme_minimal() +
  theme(legend.position = "none",
        axis.line = element_line(colour = "purple1", 
                      size = 0.5, 
                      linetype = "solid"),
        axis.title.x = element_text(size = 9),
        plot.title = element_text(size = 12, 
                                  hjust = 0.35, 
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0),
                                  color = "purple1"),
        plot.subtitle = element_text(hjust = 0.5),
        plot.margin=grid::unit(c(0,0,0,0), "mm")) +
  labs(title = "Cord Blood",
       y = "",
       x = "",
       caption = "")

ggsave("../results/plots/logfc_of_30_genes_affected_by_smoking_cord.png")

p_fc_mom <- smoking_fc_maternal |>
  arrange(desc(abs(log2fc))) |> 
  slice_head(n=30) |> 
  ggplot(aes(x=fct_reorder(Gene, abs(log2fc)), y=log2fc)) +
  geom_bar(stat="identity", fill  = "darkorange1", alpha = 0.6) + 
  coord_flip() +
  theme_minimal() +
  theme(legend.position = "none",
        axis.line = element_line(colour = "darkorange1", 
                      size = 0.5, 
                      linetype = "solid"),
        axis.title.x = element_text(size = 9),
        plot.title = element_text(size = 12, 
                                  hjust = 0.35, 
                                  face = "bold.italic", 
                                  margin = margin(0,0,0,0),
                                  color = "darkorange1"),
        plot.subtitle = element_text(hjust = 0.5),
        plot.margin=grid::unit(c(0,0,0,0), "mm")) +
  labs(title = "Maternal Blood",
       y = "",
       x = "",
       caption = "")
ggsave("../results/plots/logfc_of_30_genes_affected_by_smoking_maternal.png")


combined_top30 <- p_fc_mom + p_fc_placenta + p_fc_cord + 
     plot_annotation(title = "The 30 Genes Most Affected By Smoking",
                     caption = "Data from NCBI (2011)",
                     theme = theme(plot.title = element_text(size = 18,
                                                             hjust = 0.5, 
                                                             face = "bold", 
                                                             margin = margin(5,0,5,0),
                                                             color = "black"),
                                   plot.margin=grid::unit(c(0,0,0,0), "mm"))) +
  plot_layout(ncol = 3, widths = unit(c(3.4,3.4,3.4), c('cm', 'cm','cm')))
ggsave("../results/plots/combined_top30.png")
combined_top30


```

### Gene Set Enrichment Analysis

```{r}

# get gene sets for biological process ontology with respective genes

BP_df = msigdbr(species = "S. cerevisiae", category = "C5", subcategory = "BP")

gene_set_names <- BP_df |> 
  pull(gs_name)
BP_list <- BP_df |> 
  pull(human_gene_symbol) |>
  split(gene_set_names) 

selected_genes <- smoking_fc_placenta |>
  drop_na() |> 
  pull(log2fc) 
names(selected_genes) <- smoking_fc_placenta |>
  drop_na() |> 
  pull(Gene)

fgseaResult <- fgsea(pathways = BP_list, stats = selected_genes)

fgsea_result_sliced <- fgseaResult |> 
  arrange(NES) |> 
  slice_head(n = 15) |> 
  select(pathway, NES)

write.csv(fgsea_result_sliced, "../results/plots/fgsea_results.csv")
```

### PCR plot of gene expression in placenta tissue

```{r}

genes_large_fold_change <- smoking_fc_placenta |>
  arrange(desc(abs(log2fc))) |> 
  slice_head(n=25) |> 
  pull(Gene)

PCA <- placenta_joined |> 
  select(genes_large_fold_change) |> 
  prcomp() 

# summary(PCA)

PCA |> 
  augment(placenta_joined) |> 
  ggplot(aes(.fittedPC1, .fittedPC2, color = smoking_status)) + 
  geom_point(size = 2.5) +
  labs(title = "Placental PCA plot",
       subtitle = "Principal components are based on the top 25 genes with the largest fold changes",
       x = "PC-1",
       y = "PC-2",
       color = "Smoking Status",
       caption = "Data from NCBI (2011)") +
  theme_minimal() +
  geom_hline(yintercept=0, 
             linetype = "dashed",
             alpha = .2) +
  geom_vline(xintercept=0, 
             linetype = "dashed",
             alpha = .2) +
  theme(legend.position = c(0.8, 0.8),
        legend.direction = "vertical",
        legend.title = element_text(size=9),
        legend.text = element_text(size=8),
        legend.background = element_rect(linetype = 1, 
                                         size = 0.5, 
                                         colour = 1),
        plot.title = element_text(size = 16, 
                                  hjust = 0.5, 
                                  face = "bold", 
                                  margin = margin(10,-10,10,10),
                                  color = "darkblue"),
        plot.subtitle = element_text(size = 12,
                                     face = "italic")) +
  scale_colour_manual(values = c("yellow3","red"))

ggsave("../results/plots/pca_plot_smoking_gene_expr_placenta.png")
```

25 genes with the highest fold change in the gene expression in placenta samples between smoking mothers and nonsmoking mothers were taken.

After applying principal component analysis on the gene expression of these genes, we can see that we can see a cluster of the transformed gene expression from the nonsmoking mothers in the lower bottom corner, and a looser cluster in the top left corner of the transformed gene expression from the smoking mothers.

**We can conclude that smoking affects gene expression in placenta.**
