==> R/06_analysis_2.qmd <==
### Source functions

```{r}
source("../R/99_proj_func.R")
```

### Library:

```{r}
#| eval: false 
#| echo: true
#| message: false
library("tidyverse")
library("ggridges")
library("viridis")
library("broom")
library("ggrepel")
library("viridis")
```

### loading:

```{r}
#| eval: true
#| message: false

ncbi_joined <- vroom::vroom("../data/04_ncbi_joined.tsv.gz",
                             delim = "\t")

placenta_joined <- vroom::vroom("../data/04_placenta_joined.tsv.gz",
                                delim = "\t")

maternal_joined <- vroom::vroom("../data/04_maternal_joined.tsv.gz",
                                delim = "\t")

cord_joined <- vroom::vroom("../data/04_cord_joined.tsv.gz",
                                delim = "\t")

smoking_fc_placenta <- vroom::vroom("../data/04_smoking_fc_placenta.tsv.gz",
                                 delim = "\t")

smoking_fc_maternal <- vroom::vroom("../data/04_smoking_fc_maternal.tsv.gz",
                                 delim = "\t")

smoking_fc_cord <- vroom::vroom("../data/04_smoking_fc_cord.tsv.gz",
                                 delim = "\t")

```

### plotting:

```{r}
#| eval: false 
#| echo: true

## Magnus

# SKRALD
# SKRALD
# SKRALD

ggplot(data = ncbi_joined,
       mapping = aes(x = newborn_weight,
                     y = log2(cotinine_conc_maternal))) +
  geom_point(aes(color = age_group,
                 size = factor(BMI_class,
                               levels = c("Severely underweight",
                                          "Underweight",
                                          "Normal weight", 
                                          "Overweight",
                                          "Obesity", 
                                          "Obesity class I",
                                          "Obesity class II", 
                                          "Obesity class III")))) + 
  labs(title = "Comparision of smoking severity and newborn weight",
       x = "Newborn Weight (g)",
       y = "Log2 of cotinine_conc_maternal",
       color = "Age Group",
       size = "BMI class",
       caption = "Data from NCBI (2011)") +
  theme_minimal(base_family = "Avenir") + 
  theme(legend.position = "right") 

# GULD
# GULD
# GULD
ggplot(data = ncbi_joined,
       mapping = aes(x = smoking_status,
                     y = newborn_weight,
                     fill = smoking_status)) +
  geom_boxplot() +
  labs(title = "Investigating Effect of Smoking on Newborn",
       subtitle = "Comparision of newborn weight and smoking status",
       x = "",
       fill = "",
       y = "Newborn Weight (g)",
       caption = "Data from NCBI (2011)") +
  theme_minimal(base_family = "Avenir") + 
  theme(legend.position = "none") 


# GULD
# GULD 
# GULD
ggplot(data = ncbi_joined,
       mapping = aes(x = newborn_weight,
                     y = placental_weight)) +
  geom_point(aes(color = smoking_status,
                 size = factor(BMI_class,
                               levels = c("Severely underweight",
                                          "Underweight",
                                          "Normal weight", 
                                          "Overweight",
                                          "Obesity", 
                                          "Obesity class I",
                                          "Obesity class II", 
                                          "Obesity class III")))) + 
  geom_hline(aes(yintercept = mean(placental_weight)),
             alpha = 0.4,
             linetype="dashed",
             color = "purple", 
             linewidth=0.5) +
  geom_vline(aes(xintercept = mean(newborn_weight)),
             alpha = 0.4,
             linetype="dashed",
             color = "purple", 
             linewidth=0.5) +
  labs(title = "Weight comparison",
       subtitle = "Measured in grams",
       x = "Newborn",
       y = "Placental",
       color = "Smoking Status",
       size = "BMI class",
       caption = "Data from NCBI (2011)") +
  theme_minimal(base_family = "Avenir") + 
  theme(legend.position = "right",
        legend.direction = "vertical") 


### GULD
ggplot(data = ncbi_joined,
       mapping = aes(x = log2(cotinine_conc_maternal),
                     y = log2(cotinine_conc_cord))) +
  geom_point() +
  geom_smooth(method=lm) + 
  labs(title = "Concentration of Cotinine",
       subtitle = "Comparision of log2 transformed cotinine concentration [ng/ml]",
       x = "Maternal Cotinine",
       y = "Cord Cotinine",
       caption = "Data from NCBI (2011)") +
  theme_minimal(base_family = "Avenir") + 
  theme(legend.position = "bottom") +
  facet_wrap(~factor(BMI_class,
                               levels = c("Severely underweight",
                                          "Underweight",
                                          "Normal weight", 
                                          "Overweight",
                                          "Obesity", 
                                          "Obesity class I",
                                          "Obesity class II", 
                                          "Obesity class III")))
### GULD
ggplot(data = ncbi_joined,
       mapping = aes(x = newborn_weight,
                     y = placental_weight)) +
  geom_point(aes(
    color = factor(BMI_class,
                   levels = c("Severely underweight",
                              "Underweight",
                              "Normal weight", 
                              "Overweight",
                              "Obesity", 
                              "Obesity class I",
                              "Obesity class II", 
                              "Obesity class III")))) +
  geom_smooth(method=lm) + 
  labs(title = "Weight distribution based on BMI of mother",
       subtitle = "Comparision of weigth [g]",
       x = "Newborn [g]",
       y = "Placental [g]",
       color = "BMI class",
       caption = "Data from NCBI (2011)") +
  theme_classic(base_family = "Avenir") + 
  theme(legend.position = "right",
        legend.direction = "vertical") +
  facet_wrap(~smoking_status,
             scales = "fixed",
             as.table = TRUE)

```
### PCA

```{r}
# Fitting
pca_fit <- ncbi_joined |>
  select(where(is.numeric)) |>
  prcomp(scale = TRUE)

# Adding PCA data to joined augmented data frame
pca_fit_augment <- pca_fit |> 
  augment(ncbi_joined)
```


```{r}
# Plot PC1 vs PC2
pca_fit_augment  |> 
  ggplot(aes(x = .fittedPC1, 
             y = .fittedPC2,
             color = tissue)) + 
  geom_point() +
  theme_minimal() +
  labs(x = "PC1",
       y = "PC2",
       color = "Tissue",
       title = "Scatter plot of fitted PC1 and PC2") +
  scale_colour_brewer(palette = "Set1") +
  geom_hline(yintercept = 0, 
             linetype = "dashed",
             alpha = .2) + 
  geom_vline(xintercept = 0, 
             linetype = "dashed",
             alpha = .2)

#ggsave("../results/plots/scatter_pc1_pc2.png", 
#       plot = scatter_pc1_pc2)
#show(scatter_pc1_pc2)
```


```{r}
# Plot PC2 vs PC3
pca_fit_augment  |> 
  ggplot(aes(x = .fittedPC2, 
             y = .fittedPC3,
             color = tissue)) + 
  geom_point() +
  theme_minimal() +
  labs(x = "PC2",
       y = "PC3",
       color = "Tissue",
       title = "Scatter plot of fitted PC2 and PC3") +
  scale_colour_brewer(palette = "Set1") +
  geom_hline(yintercept = 0, 
             linetype = "dashed",
             alpha = .2) + 
  geom_vline(xintercept = 0, 
             linetype = "dashed",
             alpha = .2)

#ggsave("../results/plots/scatter_pc1_pc2.png", 
#       plot = scatter_pc1_pc2)
#show(scatter_pc1_pc2)
```
